+++ config.php
@@ -3,3 +3,5 @@ $CFG['mysql_host']='localhost';
 $CFG['mysql_dbname']='cs06';
 $CFG['mysql_username']='cs06';
 $CFG['mysql_password']='Hbds43RV5';
+$CFG['admin']='admin';
+$CFG['admin_pw']='hello';